Rb create topic

+++ HogwartsRunningClub/Controllers/TopicsController.cs
@@ -7,6 +7,8 @@ using Microsoft.AspNetCore.Mvc.Rendering;
 using Microsoft.EntityFrameworkCore;
 using HogwartsRunningClub.Data;
 using HogwartsRunningClub.Models;
+using HogwartsRunningClub.Models.ViewModels.TopicViewModels;
+using Microsoft.AspNetCore.Identity;
 
 namespace HogwartsRunningClub.Controllers
 {
@@ -14,10 +16,15 @@ namespace HogwartsRunningClub.Controllers
     {
         private readonly ApplicationDbContext _context;
 
-        public TopicsController(ApplicationDbContext context)
+        private readonly UserManager<ApplicationUser> _userManager;
+
+        public TopicsController(ApplicationDbContext context, UserManager<ApplicationUser> userManager)
         {
             _context = context;
+            _userManager = userManager;
         }
+
+        private Task<ApplicationUser> GetCurrentUserAsync() => _userManager.GetUserAsync(HttpContext.User);
 
         // GET: Topics
         public async Task<IActionResult> Index()
@@ -34,24 +41,60 @@ namespace HogwartsRunningClub.Controllers
                 return NotFound();
             }
 
-            var topic = await _context.Topic
+            Topic topic = await _context.Topic
                 .Include(t => t.TopicCategory)
                 .Include(t => t.User)
                 .FirstOrDefaultAsync(m => m.TopicId == id);
+
+            ApplicationUser user = await GetCurrentUserAsync();
+
             if (topic == null)
             {
                 return NotFound();
             }
 
-            return View(topic);
+            if (topic.HouseExclusive == true && topic.User.HouseId != user.HouseId) 
+            {
+                return RedirectToAction("ViewGreatHall", "Home");
+            }
+
+            DetailsTopicViewModel viewmodel = new DetailsTopicViewModel();
+            viewmodel.Edit = true;
+
+            if (topic.UserId != user.Id) 
+            {
+                topic.TotalViews++;
+                viewmodel.Edit = false;
+                
+            }
+
+            viewmodel.Topic = topic;
+            
+            _context.Update(topic);
+            await _context.SaveChangesAsync();
+
+            return View(viewmodel);
         }
 
         // GET: Topics/Create
-        public IActionResult Create()
+        public async Task<IActionResult> Create()
         {
-            ViewData["TopicCategoryId"] = new SelectList(_context.TopicCategory, "TopicCategoryId", "Label");
-            ViewData["UserId"] = new SelectList(_context.ApplicationUser, "Id", "Id");
-            return View();
+            CreateTopicViewModel viewmodel = new CreateTopicViewModel();
+
+            List<TopicCategory> categories = await _context.TopicCategory.ToListAsync();
+
+            List<SelectListItem> categoryOptions = categories.Select(c =>
+            {
+                 return new SelectListItem
+                        {
+                            Text = c.Label,
+                            Value = c.TopicCategoryId.ToString()
+                        };
+            }).ToList();
+
+            viewmodel.CategoryOptions = categoryOptions;
+
+            return View(viewmodel);
         }
 
         // POST: Topics/Create
@@ -59,17 +102,25 @@ namespace HogwartsRunningClub.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Create([Bind("TopicId,Title,Content,UserId,DateCreated,TotalViews,HouseExclusive,TopicCategoryId")] Topic topic)
+        public async Task<IActionResult> Create(CreateTopicViewModel viewmodel)
         {
+
+            ModelState.Remove("Topic.UserId");
+            ModelState.Remove("Topic.User");
+
             if (ModelState.IsValid)
             {
-                _context.Add(topic);
+                viewmodel.Topic.TotalViews = 0;
+                viewmodel.Topic.UserId = (await GetCurrentUserAsync()).Id;
+
+                _context.Add(viewmodel.Topic);
+
                 await _context.SaveChangesAsync();
-                return RedirectToAction(nameof(Index));
+
+                return RedirectToAction("ViewGreatHall", "Home");
             }
-            ViewData["TopicCategoryId"] = new SelectList(_context.TopicCategory, "TopicCategoryId", "Label", topic.TopicCategoryId);
-            ViewData["UserId"] = new SelectList(_context.ApplicationUser, "Id", "Id", topic.UserId);
-            return View(topic);
+
+            return View(viewmodel);
         }
 
         // GET: Topics/Edit/5
@@ -80,14 +131,32 @@ namespace HogwartsRunningClub.Controllers
                 return NotFound();
             }
 
-            var topic = await _context.Topic.FindAsync(id);
+            Topic topic = await _context.Topic.FindAsync(id);
+            ApplicationUser user = await GetCurrentUserAsync();
+
+            if (topic.UserId != user.Id) 
+            {
+                return RedirectToAction("Details", new { id = topic.TopicId });
+            }
+
+            List<TopicCategory> topicCategories = await _context.TopicCategory.ToListAsync();
+
             if (topic == null)
             {
                 return NotFound();
             }
-            ViewData["TopicCategoryId"] = new SelectList(_context.TopicCategory, "TopicCategoryId", "Label", topic.TopicCategoryId);
-            ViewData["UserId"] = new SelectList(_context.ApplicationUser, "Id", "Id", topic.UserId);
-            return View(topic);
+
+            EditTopicViewModel viewmodel = new EditTopicViewModel();
+            viewmodel.Topic = topic;
+            viewmodel.CategoryOptions = topicCategories
+                    .Select(tc => new SelectListItem { 
+                        Text = tc.Label, 
+                        Value = tc.TopicCategoryId.ToString() 
+                    })
+                    .ToList();
+
+
+            return View(viewmodel);
         }
 
         // POST: Topics/Edit/5
@@ -95,23 +164,33 @@ namespace HogwartsRunningClub.Controllers
         // more details see http://go.microsoft.com/fwlink/?LinkId=317598.
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public async Task<IActionResult> Edit(int id, [Bind("TopicId,Title,Content,UserId,DateCreated,TotalViews,HouseExclusive,TopicCategoryId")] Topic topic)
+        public async Task<IActionResult> Edit(int id, EditTopicViewModel viewmodel)
         {
-            if (id != topic.TopicId)
+            if (id != viewmodel.Topic.TopicId)
             {
                 return NotFound();
             }
 
+            ModelState.Remove("Topic.User");
+            ModelState.Remove("Topic.UserId");
+
+            
             if (ModelState.IsValid)
             {
                 try
                 {
+                    Topic topic = await _context.Topic.SingleOrDefaultAsync(t => t.TopicId == id);
+                    topic.Title = viewmodel.Topic.Title;
+                    topic.Content = viewmodel.Topic.Content;
+                    topic.TopicCategoryId = viewmodel.Topic.TopicCategoryId;
+                    topic.HouseExclusive = viewmodel.Topic.HouseExclusive;
+
                     _context.Update(topic);
                     await _context.SaveChangesAsync();
                 }
                 catch (DbUpdateConcurrencyException)
                 {
-                    if (!TopicExists(topic.TopicId))
+                    if (!TopicExists(viewmodel.Topic.TopicId))
                     {
                         return NotFound();
                     }
@@ -120,11 +199,18 @@ namespace HogwartsRunningClub.Controllers
                         throw;
                     }
                 }
-                return RedirectToAction(nameof(Index));
+                return RedirectToAction("Details", new { id = viewmodel.Topic.TopicId });
             }
-            ViewData["TopicCategoryId"] = new SelectList(_context.TopicCategory, "TopicCategoryId", "Label", topic.TopicCategoryId);
-            ViewData["UserId"] = new SelectList(_context.ApplicationUser, "Id", "Id", topic.UserId);
-            return View(topic);
+
+            List<TopicCategory> topicCategories = await _context.TopicCategory.ToListAsync();
+            viewmodel.CategoryOptions = topicCategories
+                    .Select(tc => new SelectListItem
+                    {
+                        Text = tc.Label,
+                        Value = tc.TopicCategoryId.ToString()
+                    })
+                    .ToList();
+            return View(viewmodel);
         }
 
         // GET: Topics/Delete/5
@@ -135,15 +221,23 @@ namespace HogwartsRunningClub.Controllers
                 return NotFound();
             }
 
-            var topic = await _context.Topic
+            ApplicationUser user = await GetCurrentUserAsync();
+
+            Topic topic = await _context.Topic
                 .Include(t => t.TopicCategory)
                 .Include(t => t.User)
                 .FirstOrDefaultAsync(m => m.TopicId == id);
+
             if (topic == null)
             {
                 return NotFound();
             }
 
+            if (topic.UserId != user.Id) 
+            {
+                return RedirectToAction("ViewGreatHall", "Home");
+            }
+            
             return View(topic);
         }
 
@@ -152,10 +246,28 @@ namespace HogwartsRunningClub.Controllers
         [ValidateAntiForgeryToken]
         public async Task<IActionResult> DeleteConfirmed(int id)
         {
-            var topic = await _context.Topic.FindAsync(id);
+
+            ApplicationUser user = await GetCurrentUserAsync();
+            Topic topic = await _context.Topic.SingleOrDefaultAsync(t => t.TopicId == id);
+
+            if (topic.UserId != user.Id) 
+            {
+                return RedirectToAction("ViewGreatHall", "Home");
+            }
+
+            List<Comment> comments = await _context.Comment.Where(c => c.TopicId == topic.TopicId).ToListAsync();
+
+            if (comments.Count > 0) 
+            {
+                foreach (Comment comment in comments) 
+                {
+                    _context.Remove(comment);
+                }
+            }
+
             _context.Topic.Remove(topic);
             await _context.SaveChangesAsync();
-            return RedirectToAction(nameof(Index));
+            return RedirectToAction("ViewGreatHall", "Home");
         }
 
         private bool TopicExists(int id)
